// front/utils/types/coworkingsResponse.ts
interface IResponseCoworking {
  id: string | null;
  name: string | null;
  phone: string | null;
  email: string | null;
  open: string | null;
  close: string | null;
  address: string | null;
  country: string | null;
  state: string | null;
  city: string | null;
  lat: number | null;
  long: number | null;
  capacity: number | null;
  message: string | null;
  status: string | null;
  thumbnail: string | null;
}

export default IResponseCoworking